move models to folder, add init file
