+++ server/models/__init__.py
@@ -1,0 +1,5 @@
+from sqlalchemy_serializer import SerializerMixin
+from sqlalchemy.ext.associationproxy import association_proxy
+from sqlalchemy.orm import validates
+
+from config import db
