+++ server/seed.py
@@ -8,7 +8,7 @@ from faker import Faker
 
 # Local imports
 from app import app
-from models import db
+from server.models.Orderitem import db
 
 if __name__ == '__main__':
     fake = Faker()
